réparation gros porc

+++ prog_sys/App/features/LogAndState/Data/TaskData.cs
@@ -7,7 +7,7 @@ public class TaskData
     public string TargetFilePath { get; set; }
     public string State { get; set; }
     public int TotalFilesToCopy { get; set; }
-    public int TotalFilesSize { get; set; }
+    public long TotalFilesSize { get; set; }
     public int NbFilesLeftToDo { get; set; }
     public int Progression { get; set; }
     public string Type { get; set; }
